[FIX] Implementação de endpoint CadastrarOuAtualizar Usuario parse

+++ PixAPI.Repository/Entities/Usuario.cs
@@ -7,7 +7,7 @@ namespace PixAPI.Repository.Entities
         [Key]
         public long id { get; set; }
         public string nome { get; set; }
-        public byte tipoDocumento { get; set; }
+        public int tipoDocumento { get; set; }
         public long documento { get; set; }
         public string? email { get; set; }
         public long? telefone { get; set; }
